+++ src/App.tsx
@@ -1,14 +1,14 @@
 import { FC } from 'react';
-import MainPage from './pages/main/MainPage';
+import MainPage from './pages/main/main-page';
 import {BrowserRouter, Routes, Route} from 'react-router-dom';
-import { LoginPage } from './pages/login/LoginPage';
-import { FavoritesPage } from './pages/favourites/FavouritesPage';
-import OfferPage from './pages/offer/OfferPage';
-import NotFoundPage from './pages/not-found/NotFoundPage';
-import Private from './components/Private';
+import { LoginPage } from './pages/login/login-page';
+import { FavoritesPage } from './pages/favourites/favourites-page';
+import OfferPage from './pages/offer/offer-page';
+import NotFoundPage from './pages/not-found/not-found-page';
 import { FAVOURITES_URL, LOGIN_URL, MAIN_URL, OFFER_URL } from './const/url';
 import { fetchOffersList, fetchUser } from './state/actions';
 import { useAppDispatch } from './state';
+import Private from './components/private';
 
 export interface AppProps {
 }
