+++ src/components/Private.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react';
 import { Navigate } from 'react-router-dom';
-import { AuthStatus } from '../types/authStatus';
+import { AuthStatus } from '../types/auth-status';
 import { useSelector } from 'react-redux';
 import { selectAuthStatus } from '../state/selectors';
 
@@ -11,7 +11,7 @@ export interface PrivateProps {
 
 const Private: FC<PrivateProps> = ({toUrl, children}) => {
   const authStatus = useSelector(selectAuthStatus);
-  if(authStatus === AuthStatus.NOT_AUTORIZED) {
+  if(authStatus === AuthStatus.NotAuthorized) {
     return <Navigate to={toUrl} />;
   }
   return children;
