+++ src/components/star-input.tsx
