+++ src/pages/favourites/components/favourites-offer.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react';
 import { Link } from 'react-router-dom';
 import { OFFER_URL } from '../../../const/url';
 import { Offer } from '../../../types/offer';
-import FavouriteButton from '../../../components/FavouriteButton';
+import FavouriteButton from '../../../components/favourite-button';
 import React from 'react';
