+++ src/pages/login/login-page.tsx
@@ -1,12 +1,12 @@
 import { FC, FormEvent, useState } from 'react';
-import Header from '../../components/Header';
+import Header from '../../components/header';
 import { logIn } from '../../state/actions';
 import { useAppDispatch } from '../../state';
 import { MAIN_URL } from '../../const/url';
 import { Navigate, useNavigate } from 'react-router-dom';
 import { useSelector } from 'react-redux';
 import { selectAuthStatus } from '../../state/selectors';
-import { AuthStatus } from '../../types/authStatus';
+import { AuthStatus } from '../../types/auth-status';
 
 interface FormState {
   email: string;
@@ -43,7 +43,7 @@ export const LoginPage: FC = () => {
       setFormDisabled(false);
     });
   };
-  if(authStatus === AuthStatus.AUTHORIZED) {
+  if(authStatus === AuthStatus.Authorized) {
     return <Navigate to={MAIN_URL} />;
   }
   return (
