+++ src/pages/offer/components/comment-card.tsx
