+++ src/types/auth-status.ts
@@ -1,0 +1,4 @@
+export enum AuthStatus {
+    Authorized,
+    NotAuthorized
+}
