--- src/types/authStatus.ts
@@ -1,4 +1,0 @@
-export enum AuthStatus {
-    AUTHORIZED,
-    NOT_AUTORIZED
-}
