+++ src/types/favourite-status.ts
