+++ src/types/full-offer-info.ts
